# docs/swagger.yaml
basePath: /api
definitions:
  app.AccWError:
    properties:
      error:
        type: string
    type: object
...
  instance.AccwebGlobalBanEntryJson:
    properties:
      playerId:
        type: string
      playerName:
        type: string
    type: object
  instance.AccwebGlobalBanlistJson:
    properties:
      enabled:
        type: boolean
      entries:
        items:
          $ref: '#/definitions/instance.AccwebGlobalBanEntryJson'
        type: array
    type: object
  instance.AccwebGlobalEntrylistJson:
    properties:
      enabled:
        type: boolean
      entries:
        items:
          $ref: '#/definitions/instance.EntrySettings'
        type: array
    type: object
...
  instance.DriverSettings:
    properties:
      driverCategory:
        type: integer
      firstName:
        type: string
      lastName:
        type: string
      nationality:
        type: integer
      playerID:
        type: string
      shortName:
        type: string
    type: object
...
  instance.EntrySettings:
    properties:
      ballastKg:
        type: integer
      customCar:
        type: string
      defaultGridPosition:
        type: integer
      drivers:
        items:
          $ref: '#/definitions/instance.DriverSettings'
        type: array
      forcedCarModel:
        type: integer
      isServerAdmin:
        type: integer
      overrideCarModelForCustomCar:
        type: integer
      overrideDriverInfo:
        type: integer
      raceNumber:
        type: integer
      restrictor:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    name: ACCWeb project
    url: https://github.com/assetto-corsa-web/accweb
  description: ACCweb api documentation
  license:
    name: MIT
    url: https://github.com/assetto-corsa-web/accweb/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: ACCWeb API documentation
  version: "1.19"
paths:
  /configure/global-ban:
    get:
      consumes:
      - application/json
      description: List all global bans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/instance.AccwebGlobalBanlistJson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.AccWError'
      security:
      - JWT: []
      summary: List all global bans
      tags:
      - servers
    post:
      consumes:
      - application/json
      description: Save global bans
      parameters:
      - description: Global ban entry payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/instance.AccwebGlobalBanEntryJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.AccWError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.AccWError'
      security:
      - JWT: []
      summary: Save global bans
      tags:
      - servers
  /configure/global-ban/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a global ban entry by ID
      parameters:
      - description: Ban entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.AccWError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.AccWError'
      security:
      - JWT: []
      summary: Remove a global ban entry
      tags:
      - servers
  /configure/global-ban/enable-toggle:
    post:
      consumes:
      - application/json
      description: Toggle global bans enabled status
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.AccWError'
      security:
      - JWT: []
      summary: Toggle global bans enabled status
      tags:
      - servers
  /configure/global-entrylist:
    get:
      consumes:
      - application/json
      description: List all global admins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/instance.AccwebGlobalEntrylistJson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.AccWError'
      security:
      - JWT: []
      summary: List all global admins
      tags:
      - servers
    post:
      consumes:
      - application/json
      description: Save global admins
      parameters:
      - description: Global admins payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/instance.AccwebGlobalEntrylistJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.AccWError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.AccWError'
      security:
      - JWT: []
      summary: Save global admins
      tags:
      - servers
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"